chore: move files to storybook folders
- rename files
