+++ packages/service/src/index.ts
@@ -1,6 +1,6 @@
 import type { WicElement } from 'webdriver-image-comparison/dist/commands/element.interfaces.js'
 import WdioImageComparisonService from './service.js'
-import VisualLauncher from './launcher.js'
+import VisualLauncher from './storybook/launcher.js'
 import type {
     Output,
     Result,
